dead code cleanup; linux attempts
