+++ KOKTKaraokeParty/export_presets.cfg
@@ -86,16 +86,16 @@ encryption_exclude_filters=""
 seed=0
 encrypt_pck=false
 encrypt_directory=false
-script_export_mode=2
+script_export_mode=0
 
 [preset.1.options]
 
 custom_template/debug=""
 custom_template/release=""
 debug/export_console_wrapper=1
-binary_format/embed_pck=false
+binary_format/embed_pck=true
 texture_format/s3tc_bptc=true
-texture_format/etc2_astc=false
+texture_format/etc2_astc=true
 binary_format/architecture="x86_64"
 ssh_remote_deploy/enabled=false
 ssh_remote_deploy/host="user@host_ip"
@@ -109,7 +109,7 @@ unzip -o -q \"{temp_dir}/{archive_name}\" -d \"{temp_dir}\"
 ssh_remote_deploy/cleanup_script="#!/usr/bin/env bash
 kill $(pgrep -x -f \"{temp_dir}/{exe_name} {cmd_args}\")
 rm -rf \"{temp_dir}\""
-dotnet/include_scripts_content=false
+dotnet/include_scripts_content=true
 dotnet/include_debug_symbols=true
 dotnet/embed_build_outputs=false
