+++ KOKTKaraokeParty/project.godot
@@ -37,5 +37,6 @@ project/solution_directory="../"
 
 [rendering]
 
-renderer/rendering_method="mobile"
+renderer/rendering_method="gl_compatibility"
+renderer/rendering_method.mobile="gl_compatibility"
 textures/vram_compression/import_etc2_astc=true
